requirements.txt
numpy
opencv-python
tqdm
py360convert
Pillow
transformers
safetensors
torch